requirements.txt
pycryptodome-3.9.7
django-3.0.6
Pillow